+++ scripts/build_mmg.sh
@@ -2,24 +2,28 @@
 
 pushd $SRC_DIR
 
-# download tar 
-wget https://github.com/MmgTools/mmg/archive/refs/tags/v$MMG_VERSION.tar.gz
-
-# extract source
-tar -xf v$MMG_VERSION.tar.gz
-
-# delete tar
-rm v$MMG_VERSION.tar.gz
-
-# create install directory 
 export MMG_INSTALL_DIR=$INSTALL_DIR/mmg-$MMG_VERSION
-mkdir -p $MMG_INSTALL_DIR
-echo $MMG_INSTALL_DIR
-# build and install
-pushd mmg-$MMG_VERSION
-mkdir -p build && cd build
-cmake -DCMAKE_INSTALL_PREFIX=$MMG_INSTALL_DIR ..
-make -j 4 && make install
-popd
+
+if [[ $BUILD_MMG -eq 1 ]]
+then 
+    # download tar 
+    wget https://github.com/MmgTools/mmg/archive/refs/tags/v$MMG_VERSION.tar.gz
+
+    # extract source
+    tar -xf v$MMG_VERSION.tar.gz
+
+    # delete tar
+    rm v$MMG_VERSION.tar.gz
+
+    # create install directory 
+    mkdir -p $MMG_INSTALL_DIR
+    echo $MMG_INSTALL_DIR
+    # build and install
+    pushd mmg-$MMG_VERSION
+    mkdir -p build && cd build
+    cmake -DCMAKE_INSTALL_PREFIX=$MMG_INSTALL_DIR ..
+    make -j 4 && make install
+    popd
+fi
 
 popd
